# docs/openapi.yaml
openapi: 3.1.0
info:
  title: Yasser AI Academy API
  version: 1.0.0
  description: Authenticated education, commerce, portal, tutor, and credential APIs.
servers:
  - url: http://localhost:3001/api
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: yaa_session
...
paths:
...
  /checkout:
    post:
      security: [{ sessionCookie: [] }]
      summary: Create an idempotent order
      responses:
        "201": { description: Order created }
  /payments/sandbox/{orderId}:
    post:
      security: [{ sessionCookie: [] }]
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Payment and enrollment completed }
  /account/orders:
    get:
      security: [{ sessionCookie: [] }]
      summary: List the authenticated customer's orders
      responses:
        "200": { description: Order history }
  /account/orders/{orderId}/invoice:
    get:
      security: [{ sessionCookie: [] }]
      summary: Get an invoice for an owned paid order
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "200": { description: Invoice details }
        "404": { description: Invoice not found }
  /finance/orders/{orderId}/refunds:
    post:
      security: [{ sessionCookie: [] }]
      summary: Create an idempotent full or partial refund
      parameters:
        - in: path
          name: orderId
          required: true
          schema: { type: string, format: uuid }
      responses:
        "201": { description: Refund created }
        "403": { description: Finance permission required }
        "502": { description: Configured payment provider unavailable }
...